+++ lib/templates/bootstrap_horizontal.ex
@@ -11,17 +11,18 @@ defmodule Formex.Template.BootstrapHorizontal do
     * `right_column` - left column class, defaults to `col-sm-10`
   """
 
-  def generate_row(form, field, options \\ []) do
+  def generate_row(form, item, options \\ [])
 
-    left_column   = Keyword.get(options, :left_column, "col-sm-2")
-    right_column  = Keyword.get(options, :right_column, "col-sm-10")
+  def generate_row(form, field = %Field{}, options) do
 
-    field_html  = generate_field_html(form, field)
-    label_html  = generate_label_html(form, field, left_column)
+    {left_column, right_column} = get_columns(options)
 
-    field_html = attach_addon(field_html, field)
+    input  = generate_input(form, field)
+    label  = generate_label(form, field, left_column)
 
-    tags = [field_html]
+    input = attach_addon(input, field)
+
+    tags = [input]
     |> attach_error(form, field)
 
     wrapper_class = ["form-group"]
@@ -30,8 +31,28 @@ defmodule Formex.Template.BootstrapHorizontal do
 
     column = content_tag(:div, tags, class: right_column)
 
-    content_tag(:div, [label_html, column], class: Enum.join(wrapper_class, " "))
+    content_tag(:div, [label, column], class: Enum.join(wrapper_class, " "))
 
+  end
+
+  def generate_row(form, button = %Button{}, options) do
+
+    {left_column, right_column} = get_columns(options)
+
+    offset = String.replace(left_column, ~r/([0-9]+)$/, "offset-\\1")
+    right_column = right_column<>" "<>offset
+
+    input        = generate_input(form, button)
+    input_column = content_tag(:div, input, class: right_column)
+
+    content_tag(:div, input_column, class: "form-group")
+  end
+
+  defp get_columns(options) do
+    {
+      Keyword.get(options, :left_column, "col-sm-2"),
+      Keyword.get(options, :right_column, "col-sm-10")
+    }
   end
 
 end